requirements.txt
faiss==1.7.2
ipython==7.31.1
numpy==1.21.5
Pillow==9.4.0
Pillow==10.4.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
wandb==0.15.2
